重構 h4_wikidot_rebuild 之2, 整併部份函式進 lib/Wikidot.py
將 php_parse_wikidot_heading(), list_headings() 整合進 lib/Wikidot.py
持續縮減 h4_wikidot_rebuild 的程式
